+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [0.1.2.20230118]
+
+### Add
+
+- [core/Session.php] : Implementada classe de gerenciamento de sessão.
+
+### Changed
+
+- [core/Config/Application.php] : Adicionada propriedade de timezone para armazenar a timezone da aplicação.
+- [core/Neon.php] : Adicionada chamada do método 'timezone()'.
+
+
+### Fixed
+
+- [core/Http/Response.php] : Removida função "mb_convert_encoding" para renderização de página.
+
 ## [0.1.1.20230110]
 
 ### Changed